+++ src/test/java/org/spdx/library/model/compat/v2/ChecksumTest.java
@@ -30,7 +30,7 @@ import org.spdx.library.model.v2.SpdxConstantsCompatV2;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.Version;
 import org.spdx.library.model.v2.enumerations.ChecksumAlgorithm;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 
 import junit.framework.TestCase;
