+++ src/test/java/org/spdx/library/model/compat/v2/ExternalDocumentRefTest.java
@@ -34,7 +34,7 @@ import org.spdx.library.model.v2.SpdxModelFactoryCompatV2;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.Version;
 import org.spdx.library.model.v2.enumerations.ChecksumAlgorithm;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 
 import junit.framework.TestCase;
