+++ src/test/java/org/spdx/library/model/compat/v2/ExternalRefTest.java
@@ -31,7 +31,7 @@ import org.spdx.library.model.v2.ReferenceType;
 import org.spdx.library.model.v2.SpdxConstantsCompatV2;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.enumerations.ReferenceCategory;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.library.referencetype.ListedReferenceTypes;
 import org.spdx.spdxRdfStore.RdfStore;
 import org.spdx.storage.IModelStore;
