+++ src/test/java/org/spdx/library/model/compat/v2/LineCharPointerTest.java
@@ -30,7 +30,7 @@ import org.spdx.library.model.v2.GenericSpdxElement;
 import org.spdx.library.model.v2.SpdxElement;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.pointer.LineCharPointer;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 
 import junit.framework.TestCase;
