+++ src/test/java/org/spdx/library/model/compat/v2/ModelCollectionTest.java
@@ -13,7 +13,7 @@ import org.spdx.library.ModelCopyManager;
 import org.spdx.library.model.v2.GenericSpdxElement;
 import org.spdx.library.model.v2.SpdxConstantsCompatV2;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 import org.spdx.storage.PropertyDescriptor;
