+++ src/test/java/org/spdx/library/model/compat/v2/SpdxConstantElementTest.java
@@ -30,7 +30,7 @@ import org.spdx.library.model.v2.SpdxConstantElement;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.SpdxNoAssertionElement;
 import org.spdx.library.model.v2.SpdxNoneElement;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 
 import junit.framework.TestCase;
