+++ src/test/java/org/spdx/library/model/compat/v2/SpdxNoneElementTest.java
@@ -27,7 +27,7 @@ import org.spdx.library.model.v2.SpdxElement;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.SpdxNoneElement;
 import org.spdx.library.model.v2.enumerations.RelationshipType;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 import org.spdx.storage.IModelStore;
