+++ src/test/java/org/spdx/library/model/compat/v2/SpdxPackageVerificationCodeTest.java
@@ -28,7 +28,7 @@ import org.spdx.core.ModelRegistry;
 import org.spdx.library.ModelCopyManager;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.SpdxPackageVerificationCode;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 import org.spdx.storage.IModelStore;
 import org.spdx.storage.IModelStore.IdType;
