+++ src/test/java/org/spdx/library/model/compat/v2/StartEndPointerTest.java
@@ -32,7 +32,7 @@ import org.spdx.library.model.v2.pointer.ByteOffsetPointer;
 import org.spdx.library.model.v2.pointer.LineCharPointer;
 import org.spdx.library.model.v2.pointer.SinglePointer;
 import org.spdx.library.model.v2.pointer.StartEndPointer;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.spdxRdfStore.RdfStore;
 
 import junit.framework.TestCase;
