+++ src/test/java/org/spdx/spdxRdfStore/RdfSpdxDocumentModelManagerTest.java
@@ -20,7 +20,7 @@ import org.spdx.core.SpdxInvalidIdException;
 import org.spdx.core.TypedValue;
 import org.spdx.library.model.v2.SpdxConstantsCompatV2;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.storage.CompatibleModelStoreWrapper;
 import org.spdx.storage.IModelStore.IdType;
