+++ src/test/java/org/spdx/spdxRdfStore/RdfStoreTest.java
@@ -30,7 +30,7 @@ import org.spdx.library.model.v2.SpdxModelFactoryCompatV2;
 import org.spdx.library.model.v2.SpdxModelInfoV2_X;
 import org.spdx.library.model.v2.SpdxPackage;
 import org.spdx.library.model.v2.enumerations.RelationshipType;
-import org.spdx.library.model.v3.SpdxModelInfoV3_0;
+import org.spdx.library.model.v3_0_0.SpdxModelInfoV3_0;
 import org.spdx.storage.CompatibleModelStoreWrapper;
 
 import junit.framework.TestCase;
